fix(builtins): refactor test code
