+++ common/Cargo.toml
@@ -9,7 +9,7 @@ edition = "2021"
 anyhow = "1.0.70"
 blake2b_simd = { version = "1.0.0" }
 frc42_dispatch = "3.1.0"
-fvm_ipld_encoding = { git = "https://github.com/filecoin-project/ref-fvm.git", version = "0.3.3" }
+fvm_ipld_encoding = { version = "0.3.3" }
 serde = { version = "1.0.136", features = ["derive"] }
 serde_tuple = { version = "0.5.0" }
 thiserror = "1.0.40"
